# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, draft, archived]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /courses/{courseId}/modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List modules for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Modules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a module
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleBody"
      responses:
        "201":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
...
  /lessons/{id}:
    put:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
        status:
          type: string
          enum: [active, draft, archived]
        thumbnailUrl:
          type: string
          nullable: true
        studentCount:
          type: integer
        moduleCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, price, status, studentCount, moduleCount, createdAt]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            modules:
              type: array
              items:
                $ref: "#/components/schemas/Module"
          required: [modules]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
        status:
          type: string
          enum: [active, draft, archived]
        thumbnailUrl:
          type: string
          nullable: true
      required: [title, price, status]
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        titleAr:
          type: string
          nullable: true
        order:
          type: integer
        lessonCount:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required: [id, courseId, title, order, lessonCount]
    CreateModuleBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
          nullable: true
        order:
          type: integer
      required: [title, order]
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        titleAr:
          type: string
          nullable: true
        type:
          type: string
          enum: [video, pdf, text]
        videoUrl:
          type: string
          nullable: true
        pdfUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        order:
          type: integer
      required: [id, moduleId, title, type, order]
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
          nullable: true
        type:
          type: string
          enum: [video, pdf, text]
        videoUrl:
          type: string
          nullable: true
        pdfUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        order:
          type: integer
      required: [title, type, order]
...