# lib/api-zod/src/generated/types/courseDetail.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const CourseDetail = zod
  .object({
    id: zod.number(),
    title: zod.string(),
    titleAr: zod.string().nullish(),
    description: zod.string().nullish(),
    price: zod.number(),
    status: zod.enum(["active", "draft", "archived"]),
    thumbnailUrl: zod.string().nullish(),
    studentCount: zod.number(),
    moduleCount: zod.number(),
    createdAt: zod.string(),
  })
  .and(
    zod.object({
      modules: zod.array(
        zod.object({
          id: zod.number(),
          courseId: zod.number(),
          title: zod.string(),
          titleAr: zod.string().nullish(),
          order: zod.number(),
          lessonCount: zod.number(),
          lessons: zod
            .array(
              zod.object({
                id: zod.number(),
                moduleId: zod.number(),
                title: zod.string(),
                titleAr: zod.string().nullish(),
                type: zod.enum(["video", "pdf", "text"]),
                videoUrl: zod.string().nullish(),
                pdfUrl: zod.string().nullish(),
                content: zod.string().nullish(),
                duration: zod.number().nullish(),
                order: zod.number(),
              }),
            )
            .optional(),
        }),
      ),
    }),
  );

export type CourseDetail = zod.input<typeof CourseDetail>;